// src/app/app.type/customer.type.ts
export interface customer {
  birthday: string | null;
  customerId: number | null;
  customerName: string | null;
  email: string | null;
  gender: boolean;
  numberPhone: string | null;
  password: string | null;
  rankId: number | null;
  rankName: string | null;
  totalMoney: number | null;
  totalPoints: number | null;
  urlAvatar: string | null;
  userName: string | null;
}